+++ frontend/openchat-agent/src/services/online/online.client.ts
@@ -2,13 +2,12 @@ import type { HttpAgent, Identity } from "@dfinity/agent";
 import { MsgpackCanisterAgent } from "../canisterAgent/msgpack";
 import { principalStringToBytes } from "../../utils/mapping";
 import { lastOnlineResponse, markAsOnlineResponse } from "./mappers";
+import type { MinutesOnline } from "openchat-shared";
 import {
     Empty,
     OnlineUsersLastOnlineArgs,
     OnlineUsersLastOnlineResponse,
     OnlineUsersMarkAsOnlineResponse,
-    OnlineUsersMinutesOnlineArgs,
-    OnlineUsersMinutesOnlineResponse,
 } from "../../typebox";
 
 export class OnlineClient extends MsgpackCanisterAgent {
@@ -29,23 +28,13 @@ export class OnlineClient extends MsgpackCanisterAgent {
         );
     }
 
-    markAsOnline(): Promise<number> {
+    markAsOnline(): Promise<MinutesOnline> {
         return this.executeMsgpackUpdate(
             "mark_as_online",
             {},
             markAsOnlineResponse,
             Empty,
             OnlineUsersMarkAsOnlineResponse,
-        );
-    }
-
-    minutesOnline(year: number, month: number): Promise<number> {
-        return this.executeMsgpackQuery(
-            "minutes_online",
-            { year, month },
-            (resp) => resp.Success,
-            OnlineUsersMinutesOnlineArgs,
-            OnlineUsersMinutesOnlineResponse,
         );
     }
 }