+++ frontend/openchat-shared/src/domain/worker.ts
@@ -220,6 +220,7 @@ import type {
     UserSummary,
 } from "./user";
 import type { Verification } from "./wallet";
+import type { MinutesOnline } from "./online";
 
 /**
  * Worker request types
@@ -277,7 +278,6 @@ export type WorkerRequest =
     | ChatEventsWindow
     | LastOnline
     | MarkAsOnline
-    | MinutesOnline
     | GetGroupDetails
     | MarkMessagesRead
     | GetAllCachedUsers
@@ -1268,12 +1268,6 @@ type MarkAsOnline = {
     kind: "markAsOnline";
 };
 
-type MinutesOnline = {
-    year: number;
-    month: number;
-    kind: "minutesOnline";
-};
-
 type FreezeGroup = {
     chatId: GroupChatIdentifier;
     reason: string | undefined;
@@ -1330,6 +1324,8 @@ type RemoveMessageFilter = {
 type SetAirdropConfig = {
     channelId: number;
     channelName: string;
+    communityId?: string;
+    communityName?: string;
     kind: "setAirdropConfig";
 };
 
@@ -2097,9 +2093,7 @@ export type WorkerResult<T> = T extends Init
     : T extends LastOnline
     ? Record<string, number>
     : T extends MarkAsOnline
-    ? number
-    : T extends MinutesOnline
-    ? number
+    ? MinutesOnline
     : T extends ChatEventsBatch
     ? ChatEventsResponse[]
     : T extends ChatEventsWindow